+++ scripts/brew/templates/secman.rb
@@ -1,0 +1,19 @@
+class Secman < Formula
+  desc "The Password Manager of your dreams"
+  homepage "https://secman.dev/docs/cli"
+  url "__CLI_DOWNLOAD_URL__"
+  sha256 "__CLI_SHA256__"
+  depends_on "scmn-dev/secman/sm-node" => "__NODE_VERSION__"
+
+  def install
+    inreplace "bin/secman", /^CLIENT_HOME=/, "export SECMAN_OCLIF_CLIENT_HOME=#{lib/"client"}\nCLIENT_HOME="
+    inreplace "bin/secman", "\"$DIR/node\"", Formula["sm-node"].opt_bin/"node"
+    libexec.install Dir["*"]
+    bin.install_symlink libexec/"bin/secman"
+    system("export SM_PROVIDER=brew")
+  end
+
+  test do
+    system bin/"secman", "-v"
+  end
+end