+++ v1/tools/looksh/looksh.go
@@ -1,9 +1,0 @@
-// +build !windows
-
-package looksh
-
-import "os/exec"
-
-func Look() (string, error) {
-	return exec.LookPath("sh")
-}